Pipes.Concurrent

Description

Asynchronous communication between pipes

Inputs and Outputs

newtype Input a

An exhaustible source of values

recv returns Nothing if the source is exhausted

Constructors

Input

Fields

recv :: STM (Maybe a)

newtype Output a

An exhaustible sink of values

send returns False if the sink is exhausted

Constructors

Output

Fields

send :: a -> STM Bool

Pipe utilities

fromInput :: forall (m :: Type -> Type) a. MonadIO m => Input a -> Producer' a m ()

Convert an Input to a Producer

fromInput terminates when the Input is exhausted.

toOutput :: forall (m :: Type -> Type) a. MonadIO m => Output a -> Consumer' a m ()

Convert an Output to a Consumer

toOutput terminates when the Output is exhausted.

Mailboxes

type Mailbox a = (Output a, Input a)

Combines a source of values and a sink of values

fromMailbox uses Mailbox as Producer toMailbox uses Mailbox as Consumer 'send'' puts a value in the Mailbox 'recv'' obtains a value from the Mailbox

fromMailbox :: forall (m :: Type -> Type) a. MonadIO m => Mailbox a -> Producer' a m ()

Convert a Mailbox to a Producer

fromMailbox terminates when the Mailbox source of values is exhausted.

toMailbox :: forall (m :: Type -> Type) a. MonadIO m => Mailbox a -> Consumer' a m ()

Convert a Mailbox to a Consumer

toMailbox terminates when the Mailbox sink of values is exhausted.

send' :: Mailbox a -> a -> STM Bool

Put a value in a Mailbox

send returns False if the Mailbox sink is exhausted

recv' :: Mailbox a -> STM (Maybe a)

Obtain a value from a Mailbox

recv returns Nothing if the Mailbox source is exhausted

Actors

spawn :: Buffer a -> IO (Output a, Input a)

Spawn a mailbox using the specified Buffer to store messages

Using send on the Output

fails and returns False if the mailbox is sealed, otherwise it:

retries if the mailbox is full, or:

adds a message to the mailbox and returns True.

Using recv on the Input:

retrieves a message from the mailbox wrapped in Just if the mailbox is not empty, otherwise it:

retries if the mailbox is not sealed, or:

fails and returns Nothing.

If either the Input or Output is garbage collected the mailbox will become sealed.

spawn' :: Buffer a -> IO (Output a, Input a, STM ())

Like spawn, but also returns an action to manually seal the mailbox early:

(output, input, seal) <- spawn' buffer
...

Use the seal action to allow early cleanup of readers and writers to the mailbox without waiting for the next garbage collection cycle.

withSpawn :: Buffer a -> ((Output a, Input a) -> IO r) -> IO r

withSpawn passes its enclosed action an Output and Input like you'd get from spawn, but automatically seals them after the action completes. This can be used when you need the sealing behavior available from 'spawn'', but want to work at a bit higher level:

withSpawn buffer $ \(output, input) -> ...

withSpawn is exception-safe, since it uses bracket internally.

withBuffer :: Buffer a -> (Output a -> IO l) -> (Input a -> IO r) -> IO (l, r)

A more restrictive alternative to withSpawn that prevents deadlocks

data Buffer a

Buffer specifies how to buffer messages stored within the mailbox

Constructors

Unbounded

Deprecated: Use unbounded instead

Bounded Int

Deprecated: Use bounded instead

Single

Deprecated: Use bounded 1 instead

Latest a

Deprecated: Use latest instead

Newest Int

Deprecated: Use newest instead

New

Deprecated: Use newest 1 instead

unbounded :: Buffer a

Store an unbounded number of messages in a FIFO queue

bounded :: Int -> Buffer a

Store a bounded number of messages, specified by the Int argument

latest :: a -> Buffer a

Only store the Latest message, beginning with an initial value

Latest is never empty nor full.

newest :: Int -> Buffer a

Like Bounded, but send never fails (the buffer is never full). Instead, old elements are discarded to make room for new elements

Re-exports

Control.Concurrent re-exports forkIO, although I recommend using the async library instead.

Control.Concurrent.STM re-exports atomically and STM.

System.Mem re-exports performGC.

forkIO :: IO () -> IO ThreadId

Creates a new thread to run the IO computation passed as the first argument, and returns the ThreadId of the newly created thread.

The new thread will be a lightweight, unbound thread. Foreign calls made by this thread are not guaranteed to be made by any particular OS thread; if you need foreign calls to be made by a particular OS thread, then use forkOS instead.

The new thread inherits the masked state of the parent (see mask).

The newly created thread has an exception handler that discards the exceptions BlockedIndefinitelyOnMVar, BlockedIndefinitelyOnSTM, and ThreadKilled, and passes all other exceptions to the uncaught exception handler.

WARNING: Exceptions in the new thread will not be rethrown in the thread that created it. This means that you might be completely unaware of the problem if/when this happens. You may want to use the async library instead.

data STM a

A monad supporting atomic memory transactions.

atomically :: STM a -> IO a

Perform a series of STM actions atomically.

Using atomically inside an unsafePerformIO or unsafeInterleaveIO subverts some of guarantees that STM provides. It makes it possible to run a transaction inside of another transaction, depending on when the thunk is evaluated. If a nested transaction is attempted, an exception is thrown by the runtime. It is possible to safely use atomically inside unsafePerformIO or unsafeInterleaveIO, but the typechecker does not rule out programs that may attempt nested transactions, meaning that the programmer must take special care to prevent these.

However, there are functions for creating transactional variables that can always be safely called in unsafePerformIO. See: newTVarIO, newTChanIO, newBroadcastTChanIO, newTQueueIO, newTBQueueIO, and newTMVarIO.

Using unsafePerformIO inside of atomically is also dangerous but for different reasons. See unsafeIOToSTM for more on this.

newTVarIO :: a -> IO (TVar a)

IO version of newTVar. This is useful for creating top-level TVars using unsafePerformIO, because using atomically inside unsafePerformIO isn't possible.

readTVar :: TVar a -> STM a

Return the current value stored in a TVar.

mkWeakTVar :: TVar a -> IO () -> IO (Weak (TVar a))

Make a Weak pointer to a TVar, using the second argument as a finalizer to run when TVar is garbage-collected

Since: stm-2.4.3

performGC :: IO ()

Triggers an immediate major garbage collection.